# api/openapi.yaml
openapi: 3.0.3
info:
  title: Golf Course Reservation Management System API
  description: 골프장 예약 관리 시스템 API
  version: 1.0.0
  contact:
    name: API Support
    email: support@golfadmin.com
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://api.golfadmin.com
    description: Production server
security:
  - bearerAuth: []
...
paths:
...
  /tee-times/confirm:
    post:
      tags:
        - TeeTimes
      summary: 티타임 예약 확정
      operationId: confirmTeeTime
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfirmTeeTimeRequest'
      responses:
        '200':
          description: 확정 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeeTime'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
  /tee-times/cancel:
    post:
      tags:
        - TeeTimes
      summary: 티타임 예약 취소
      operationId: cancelTeeTime
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CancelTeeTimeRequest'
      responses:
        '200':
          description: 취소 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeeTime'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
...
  /tee-times/{id}:
    get:
      tags:
        - TeeTimes
      summary: 티타임 상세 조회
      operationId: getTeeTimeById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 티타임 상세 정보
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeeTime'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'

    put:
      tags:
        - TeeTimes
      summary: 티타임 수정
      operationId: updateTeeTime
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTeeTimeRequest'
      responses:
        '200':
          description: 수정 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeeTime'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'

    delete:
      tags:
        - TeeTimes
      summary: 티타임 삭제
      operationId: deleteTeeTime
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: 삭제 성공
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'

  # Performance Endpoints
...
  /teams:
    get:
      tags:
        - Teams
      summary: 팀 목록 조회
      operationId: getTeamsList
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/SortParam'
      responses:
        '200':
          description: 팀 목록
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamListResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'

    post:
      tags:
        - Teams
      summary: 팀 생성
      operationId: createTeam
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTeamRequest'
      responses:
        '201':
          description: 팀 생성 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Team'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /teams/{id}:
    get:
      tags:
        - Teams
      summary: 팀 상세 조회
      operationId: getTeamById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 팀 상세 정보
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Team'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'

    put:
      tags:
        - Teams
      summary: 팀 수정
      operationId: updateTeam
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTeamRequest'
      responses:
        '200':
          description: 수정 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Team'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'

    delete:
      tags:
        - Teams
      summary: 팀 삭제
      operationId: deleteTeam
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: 삭제 성공
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageParam:
      name: page
      in: query
      description: 페이지 번호 (1부터 시작)
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      description: 페이지당 항목 수
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortParam:
      name: sort
      in: query
      description: 정렬 기준 (field:asc 또는 field:desc)
      schema:
        type: string
        pattern: '^[a-zA-Z]+:(asc|desc)$'
  schemas:
...
    AccountType:
      type: string
      enum:
        - SUPER_ADMIN
        - ADMIN
        - TEAM_LEADER
        - INTERNAL_MANAGER
        - EXTERNAL_MANAGER
        - PARTNER
        - GOLF_COURSE
        - MEMBER
    Region:
      type: string
      enum:
        - 제주
        - 경기
        - 강원
        - 충청
        - 호남
        - 영남
        - 경상
        - 기타
    TeeTimeStatus:
      type: string
      enum:
        - AVAILABLE
        - RESERVED
        - CONFIRMED
        - COMPLETED
        - CANCELLED
    TeeTimeType:
      type: string
      enum:
        - DAILY
        - PACKAGE
    BookingType:
      type: string
      enum:
        - BOOKING
        - JOIN
    TimeSlot:
      type: string
      enum:
        - 1부
        - 2부
        - 3부
    GolfCourseStatus:
      type: string
      enum:
        - ACTIVE
        - API_INTEGRATION
        - MANUAL
        - INACTIVE

    # Error Response
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: 에러 코드
        message:
          type: string
          description: 에러 메시지
        details:
          type: object
          description: 추가 에러 상세 정보
    ValidationError:
      type: object
      required:
        - code
        - message
        - errors
      properties:
        code:
          type: string
          enum: [VALIDATION_ERROR]
        message:
          type: string
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
              code:
                type: string
...
    User:
      type: object
      properties:
        id:
          type: string
        phoneNumber:
          type: string
        name:
          type: string
        accountType:
          $ref: '#/components/schemas/AccountType'
        teamId:
          type: string
          nullable: true
        team:
          $ref: '#/components/schemas/Team'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Team:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        leaderId:
          type: string
        leader:
          $ref: '#/components/schemas/User'
        members:
          type: array
          items:
            $ref: '#/components/schemas/User'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateTeamRequest:
      type: object
      required:
        - name
        - leaderId
      properties:
        name:
          type: string
        leaderId:
          type: string
    UpdateTeamRequest:
      type: object
      properties:
        name:
          type: string
        leaderId:
          type: string
    TeamListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Team'
        pagination:
          $ref: '#/components/schemas/Pagination'

    # Golf Course Schemas
    GolfCourse:
      type: object
      properties:
        id:
          type: string
        order:
          type: integer
        region:
          $ref: '#/components/schemas/Region'
        name:
          type: string
        address:
          type: string
        phoneNumber:
          type: string
        status:
          $ref: '#/components/schemas/GolfCourseStatus'
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    TeeTime:
      type: object
      properties:
        id:
          type: string
        golfCourseId:
          type: string
        golfCourse:
          $ref: '#/components/schemas/GolfCourse'
        date:
          type: string
          format: date
        time:
          type: string
          pattern: '^([01][0-9]|2[0-3]):[0-5][0-9]$'
        timeSlot:
          $ref: '#/components/schemas/TimeSlot'
        greenFee:
          type: number
          format: decimal
          description: 그린피 (만원 단위, 소수점 1자리)
        playerCount:
          type: integer
          minimum: 1
          maximum: 4
        requirements:
          type: string
        holeType:
          type: string
          enum: [18홀, 9홀]
        caddie:
          type: boolean
        advance:
          type: number
          format: decimal
          description: 선입금
        mealIncluded:
          type: boolean
        cartIncluded:
          type: boolean
        teeTimeType:
          $ref: '#/components/schemas/TeeTimeType'
        bookingType:
          $ref: '#/components/schemas/BookingType'
        status:
          $ref: '#/components/schemas/TeeTimeStatus'
        reservedAt:
          type: string
          format: date-time
          nullable: true
        managerId:
          type: string
          nullable: true
        manager:
          $ref: '#/components/schemas/User'
        confirmedAt:
          type: string
          format: date-time
          nullable: true
        confirmedBy:
          type: string
          nullable: true
        connectedTeeTimeId:
          type: string
          nullable: true
          description: 연결된 티타임 ID (패키지용)
        accommodationInfo:
          type: string
          nullable: true
          description: 숙박 정보 (패키지용)
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    UpdateTeeTimeRequest:
      type: object
      properties:
        date:
          type: string
          format: date
        time:
          type: string
          pattern: '^([01][0-9]|2[0-3]):[0-5][0-9]$'
        greenFee:
          type: number
          format: decimal
        playerCount:
          type: integer
          minimum: 1
          maximum: 4
        requirements:
          type: string
        holeType:
          type: string
          enum: [18홀, 9홀]
        caddie:
          type: boolean
        advance:
          type: number
          format: decimal
        mealIncluded:
          type: boolean
        cartIncluded:
          type: boolean
        connectedTeeTimeId:
          type: string
        accommodationInfo:
          type: string
...
    ConfirmTeeTimeRequest:
      type: object
      required:
        - teeTimeId
      properties:
        teeTimeId:
          type: string
        notes:
          type: string
          description: 확정 메모
    CancelTeeTimeRequest:
      type: object
      required:
        - teeTimeId
      properties:
        teeTimeId:
          type: string
        reason:
          type: string
          description: 취소 사유
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
  responses:
    BadRequestError:
      description: 잘못된 요청
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
    UnauthorizedError:
      description: 인증 실패
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: UNAUTHORIZED
            message: 인증이 필요합니다
    ForbiddenError:
      description: 권한 부족
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: FORBIDDEN
            message: 해당 작업을 수행할 권한이 없습니다
    NotFoundError:
      description: 리소스를 찾을 수 없음
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: NOT_FOUND
            message: 요청한 리소스를 찾을 수 없습니다
    ConflictError:
      description: 충돌 발생
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: CONFLICT
            message: 리소스 충돌이 발생했습니다
...